- A insulina nunca fica negativa (em todas as atividades). - Nos meus dados os seguintes campos aceitam agora valores decimais: o limite para hiperglicemia, o limite para hipoglicemia, Rácio Hidratos de Carbono e Fator de sensibilidade. - Logbook completo. onTextChanged no Logbook está desativado(falta o icone apenas)
git-svn-id: https://svn.dcc.fc.up.pt/projects/pbrandao/MyDiabetes@183 8fe4487d-d89a-4624-bb73-233bb408955e

+++ src/com/jadg/mydiabetes/MyData.java
@@ -152,8 +152,8 @@ public class MyData extends Activity {
 			obj[0] = 0;
 		obj[1] = name.getText().toString();
 		obj[2] = dType.getSelectedItem().toString();
-		obj[3] = Integer.parseInt(iRatio.getText().toString());
-		obj[4] = Integer.parseInt(cRatio.getText().toString());
+		obj[3] = Double.parseDouble(iRatio.getText().toString());
+		obj[4] = Double.parseDouble(cRatio.getText().toString());
 		obj[5] = Double.parseDouble(lRange.getText().toString());
 		obj[6] = Double.parseDouble(hRange.getText().toString());
 		obj[7] = bDate.getText().toString();
